refactor(web): move misplaced files to correct directories per CONVENTIONS.md (LUM-1984)
- platform-detection.ts: utils/ → runtime/ (platform bridge, not pure utility)
- initial-message-launch.ts: utils/ → lib/ (sessionStorage side effects)
- last-viewed-conversation-storage.ts: utils/ → lib/ (localStorage side effects)
- app-pin-storage.ts: utils/ → lib/ (localStorage side effects)
- onboarding-cleanup.ts: eliminate domains/ import inversion by clearing
  localStorage keys directly instead of going through Zustand store
  (all callers navigate away immediately, so in-memory state is irrelevant)

Closes LUM-1984
