+++ apps/web/src/domains/conversations/use-conversation-loader.ts
@@ -21,7 +21,7 @@ import {
 import {
   loadLastViewedConversationId,
   saveLastViewedConversationId,
-} from "@/utils/last-viewed-conversation-storage";
+} from "@/lib/last-viewed-conversation-storage";
 import type { TranscriptPaginationState } from "@/domains/chat/transcript/types";
 import type { ContextWindowUsage } from "@/domains/chat/components/context-window-indicator";
