+++ apps/web/src/runtime/platform-detection.ts
