+++ apps/web/src/stores/pinned-apps-store.ts
@@ -18,7 +18,7 @@ import {
   pinApp,
   unpinApp,
   type PinnedAppEntry,
-} from "@/utils/app-pin-storage";
+} from "@/lib/app-pin-storage";
 
 // ---------------------------------------------------------------------------
 // Unpin event listeners
